chore: move components and hooks to src folders
