+++ note.txt
@@ -1,0 +1,11 @@
+https://github.com/gustavocd/rtkq-msw/tree/main/src
+
+https://www.freecodecamp.org/news/how-to-write-unit-tests-in-react-redux/
+
+https://gustavocd.dev/posts/testing-rtk-query-with-msw/
+
+https://github.com/vercel/next.js/tree/canary/examples/with-msw/mocks
+
+https://github.com/victoriacheng15/rest-countries-api-ts/blob/main/src/types/hooksType.d.ts
+
+https://github.com/reduxjs/redux-toolkit/issues/1271
