+++ src/app/layout.tsx
@@ -2,8 +2,8 @@
 import { useReducer } from "react";
 import { Providers } from "../redux/provider";
 import { Lato } from "next/font/google";
-import Header from "./components/Header";
-import Footer from "./components/Footer";
+import Header from "../components/Header";
+import Footer from "../components/Footer";
 import "./globals.css";
 
 const lato = Lato({
